Struct cpio_reader::Entry

pub struct Entry<'a> { /* private fields */ }

An entry of a cpio file.

Implementations

impl<'a> Entry<'a>

pub fn dev(&self) -> Option<u32>

Returns the device number of the device which contained the file.

This method returns [None] if the cpio file format is either New ASCII Format or New CRC Format. For these formats, use Entry::devmajor and Entry::devminor.

pub fn devmajor(&self) -> Option<u32>

Returns the major number of the device which contained the file.

This method returns [None] if the entry format is either Old Binary Format or Portable ASCII Format. For these formats, use Entry::dev.

pub fn devminor(&self) -> Option<u32>

Returns the minor number of the device which contained the file.

This method returns [None] if the entry format is either Old Binary Format or Portable ASCII Format. For these formats, use Entry::dev.

pub fn ino(&self) -> u32

Returns the inode number of the file.

pub fn mode(&self) -> Mode

Returns the Mode value of the file, which contains the fileโs permission information and file type.

pub fn uid(&self) -> u32

Returns the user id of the owner of the file.

pub fn gid(&self) -> u32

Returns the group id of the owner of the file.

pub fn nlink(&self) -> u32

Returns the number of links to this file.

pub fn rdev(&self) -> Option<u32>

Returns the associated device number if the entry is block special device or character special device. For the other types of entries, the caller should not use this method.

This method returns [None] if the entry format is either New ASCII Format or New CRC Format. For these formats, use Entry::rdevmajor and Entry::rdevminor.

pub fn rdevmajor(&self) -> Option<u32>

Returns the associated device major number if the entry is block special device or character special device. For the other types of entries, the caller should not use this method.

This method returns [None] if the entry format is either Old Binary Format or Portable ASCII Format. For these formats, use Entry::rdev.

pub fn rdevminor(&self) -> Option<u32>

Returns the associated device minor number if the entry is block special device or character special device. For the other types of entries, the caller should not use this method.

This method returns [None] if the entry format is either Old Binary Format or Portable ASCII Format. For these formats, use Entry::rdev.

pub fn mtime(&self) -> u64

Returns the modification time of this file.

pub fn name(&self) -> &'a str

Returns the filename.

pub fn file(&self) -> &'a [u8]

Returns the content of this file.

This method returns the path to the original file if the file is a symbolic link. For the New ASCII Format and New CRC Format, this method returns an empty slice if the file is a hard link and is not the last entry of the multiple duplicate files.